bw/requirements.txt
websocket-client
numpy
pandas
plotly
seaborn
matplotlib
